Implemented Version command
Version now reads the version number from resources/vals.properties, which is generated by maven on launch. Also commited an unfinished attempt of making the help command dynamic. Current have an issue with it reading the value as a non terminal though.. e.g. help midi thinks midi is a non terminal so it breaks.
story: 2