Trying to get scales to work correctly.
Have not succeeded yet. I have so far made a method to check for doublesharp or doubleflats so that we don't return scales for them. I also made a method that will return the enharmonic when given a letter that uses that letter. However, I want to use some sort of circular array to get one of each letter in the scale, but it's tricky so I haven't succeeded at that yet.

Story: 102