Musical Term
Set up musical term. If you type in "meaning of" followed by a musical term, it will return the definition of that word and its origin.

Currently it will only return the definition for the word lento. I need to store all of the words and their definitions in a separate file so it fetches from the file rather than the constructor, so I have started another data class to make a dictionary

story #111