Added in Minor scale keysignatures.
First I tried generating the key signatures from the scale because i thought this would work however it didn't so i went back to my old way of doing it. I added the minor key signatures and changed the displayKeySignature function to work for both major and minor notes.

story: 292