Fixed swing bug where rhythm timings were off (playing first timing twice in a row) e.g. 2:1 timing would play as 2:2:1
Still have another bug, to take notice to which fractions have to be timing divisions need to be whole numbers, and so are rounded sometimes causing inaccuracy.
This is due to all timing divisions being fractions of a beat resolution of 24. So 3:2 timing cannot accurately work this division however, as 24 cannot be cleanly divided by 5
story:291