Ensures that all commands have relevant, useful help information.
Alters some formatting, removes some unnecessary info, and implements "find chord all" as a keyword you can get help on.

Adds help info to "help", undo and redo.

Fixes the key signature help - some of it was linked to the wrong commands. My mistake.

A lot of this is minor text changes.

Story: 448