Major User Page refactor (still in progress and doesn't work).
Now have a User Page outer shell with the side menu. The summary is loaded via a seperate fxml. The tutor stats layout is also stored in a seperate fxml and has it's own controller, which is loaded from the summary controller, or straight from the userPage depending on if the summary is shown or a specific tutor.
story:620